For Study : Added this like function

+++ app/src/main/java/com/jerry/mvvm/network/ApiService.kt
@@ -1,7 +1,7 @@
 package com.jerry.mvvm.network
 
-import com.jerry.mvvm.model.ContentDetailResponse
-import com.jerry.mvvm.model.ContentListResponse
+import com.jerry.mvvm.model.remote.ContentDetailResponse
+import com.jerry.mvvm.model.remote.ContentListResponse
 import retrofit2.http.GET
 import retrofit2.http.Path
 
@@ -12,6 +12,6 @@ interface ApiService {
     suspend fun getContentList(): ContentListResponse
 
     @GET("/jerrycho/mvvm/main/json/detail_{contentId}.json")
-    suspend fun getContentById(@Path("contentId") contentId: Int): ContentDetailResponse
+    suspend fun getContentById(@Path("contentId") contentId: Long): ContentDetailResponse
 
 }
