+++ app/src/main/java/com/jerry/mvvm/repository/NetworkRepository.kt
@@ -1,12 +1,12 @@
 package com.jerry.mvvm.repository
 
 import com.jerry.mvvm.network.ApiService
-import com.jerry.mvvm.model.ContentDetailResponse
-import com.jerry.mvvm.model.ContentListResponse
+import com.jerry.mvvm.model.remote.ContentDetailResponse
+import com.jerry.mvvm.model.remote.ContentListResponse
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 
-    class NetworkRepository : KoinComponent {
+class NetworkRepository : KoinComponent {
 
     val apiService : ApiService by inject()
 
@@ -14,7 +14,7 @@ import org.koin.core.component.inject
         return  apiService.getContentList()
     }
 
-    suspend fun getContentById(id: Int): ContentDetailResponse {
+    suspend fun getContentById(id: Long): ContentDetailResponse {
         return  apiService.getContentById(id)
     }
 }
