+++ app/src/main/java/com/jerry/mvvm/ui/content/viewmodel/ContentDetailViewModel.kt
@@ -2,13 +2,11 @@ package com.jerry.mvvm.ui.content.viewmodel
 
 import androidx.lifecycle.MutableLiveData
 import com.jerry.mvvm.base.BaseViewModel
-import com.jerry.mvvm.model.ContentDetailResponse
-import com.jerry.mvvm.model.ContentListResponse
-import com.jerry.mvvm.model.ResponseContainer
+import com.jerry.mvvm.model.remote.ContentDetailResponse
+import com.jerry.mvvm.model.remote.ResponseContainer
 import com.jerry.mvvm.ui.content.usecase.ContentUseCase
 
 
-import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
 import org.koin.core.component.KoinComponent
@@ -20,7 +18,7 @@ class ContentDetailViewModel ()
     open var contentDetailMutableLiveData = MutableLiveData<ResponseContainer<ContentDetailResponse>>()
     val mContentUseCase : ContentUseCase by inject()
 
-    fun getContentById(id: Int?){
+    fun getContentById(id: Long?){
         mUiScope.launch {
             contentDetailMutableLiveData.value = ResponseContainer.loading();
             showLoadingIndicator(true)
